requirements.txt
boto==2.49.0
boto3==1.9.34
botocore==1.12.34
bz2file==0.98
certifi==2018.10.15
chardet==3.0.4
Click==7.0
docutils==0.14
gensim==3.6.0
idna==2.7
jieba==0.39
jmespath==0.9.3
joblib==0.12.5
lmdb==0.94
marisa-trie==0.7.5
mecab-python3==0.7
mwparserfromhell==0.5.2
numpy==1.15.3
pandas==0.23.4
python-dateutil==2.7.5
pytz==2018.7
requests==2.20.0
s3transfer==0.1.13
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
smart-open==1.7.1
tqdm==4.28.1
urllib3==1.24
wikipedia2vec==0.2.7
